test: move register code to bootstrap
